requirements.txt
# LBAF dependencies
vtk==9.0.1;python_version == '3.8'
vtk==9.1.0;python_version == '3.9'
numpy==1.22.3
PyYAML==6.0
Brotli==1.0.9
schema==0.7.5
scikit-learn==1.0.2
colorama==0.4.6
matplotlib==3.5.3;python_version == '3.8'
matplotlib==3.6.2;python_version == '3.9'

# LBAF testing
tox==4.6.0
coverage==6.3.2
pylint==2.12.2
pytest==7.1.1
anybadge==1.9.0

# LBAF documentation
Jinja2==3.1.2
Pygments==2.13.0
docutils==0.19

# LBAF building package
build==0.7.0
